Make lazy accessors thread-safe
This commit adds a mutex protecting each lazy attribute on first use,
in order to avoid a race condition resulting in computing the value
twice.

There are caveats related to defining accessors: for instance, it's not
thread-safe to open a class in two threads and add lazy accessors in
each thread. This use case was deemed to be too exotic to support.

This work was inspired by a question from Maxim Gurin.